Fixed multiple errors

+++ Minigin/Structs.h
@@ -1,8 +1,8 @@
 #pragma once
 struct Rect
 {
-	float x{ 0 };
-	float y{ 0 };
-	float width{ 0 };
-	float height{ 0 };
+	int x{ 0 };
+	int y{ 0 };
+	int width{ 0 };
+	int height{ 0 };
 };
